requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
python-docx==1.1.2
